move data files into data folder, change default selection for dropdown menus to Select, add item description to results
